+++ features/support/app_pages/iva/military_scra.rb
@@ -74,7 +74,7 @@ class MilitarySCRA < IDmeBase
   end
 
   def required_fields
-    [0,1,2,3,4,5,6]
+    [0,1,2,3,4,5]
   end
 
 end