+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ### Released
 
 All notable changes to this project will be documented in this file. This project does adhere to [Semantic Versioning](http://semver.org/)
+
 ## 2.2.74 (2/27/2017)
 - CORE-7034: Fixed broken Teacher Lookup tests
 - CORE-7004: Fixed broken PayPal IDP test
 - CORE-7003: Fixed broken LinkedIn IDP test
 - CORE-7018: Fixed broken SCRA tests
+- INT-1053: Shop admin ability to filter by active and inactive offers
 
 ## 2.2.73 (2/24/2017)
 - CORE-7013: Fixed broken DD214 tests